Updates app_list_dialog display to include app name.
This makes it easier to tell which bundles are which apps.  For example, the game Clash of Clans uses the bundle id com.supercell.magic, which isn't very descriptive.  Bundle Names however, are not necessarily unique (ie. 'Angry Birds' is shown for both com.rovio.angrybirdsrio and com.rovio.angrybirdsfree) so I thought it best to show both.  Ideally, these would be in different columns and sortable or something, but that's a much bigger change.