Better parsing of otool output
The output of otool is now converted into a hash where the column
headers are keys into the mach_hash. This will allow us to easily get
other values from otool in case we need them.

The newest version of otool no longer echos the path of the binary being
analyzed in the first line of the output. This broke the parsing code in
this file which should now be fixed.