Fixed SSH Connection Bug
the `connect` method was misspelled (3 n's). Also when moving the code
from the `initialize` method to `connect` the variables weren't switched
to use the instance variable versions.

Finally, none of this was being ouput because of some weird unexpected
behavior in nested exception handling code. The outermost begin rescue
in `lib/idb.rb` silently caught exceptions. This is somehow effecting
the exception handling code that was in `lib/lib/ssh_operations.rb`.